Fixes Scala 2.11 warnings on @transient
Made the @transient more precise, to tell it to apply to the generated fields.
Kept the annotation, even if a lot of them are no-op, as the fields are not generated for these
class paramaters. But it can limit bugs down the line.

Removed the anotations on the abstract element. They just don't make sense.